lazy_smp_changed.cpp
This code was forked from another source. This is just a part of that code for experimentation purpose. 
I made changes as in changing the number of nodes. I also deliberately removed some search params to observe the behaviour of the code. Since it was a vague idea to me, I couldn't present this in the blog.